pref: code structure
